Fix x64 build memory reading Move build to x64 Fix entity debouncing Added some backstops to prevent reading corrupted memory

+++ Daybreak/Models/Interop/GlobalContext.cs
@@ -7,14 +7,14 @@ namespace Daybreak.Models.Interop;
 public readonly struct GlobalContext
 {
     [FieldOffset(0x0008)]
-    public readonly IntPtr InstanceContext;
+    public readonly int InstanceContext;
 
     [FieldOffset(0x0014)]
-    public readonly IntPtr MapContext;
+    public readonly int MapContext;
 
     [FieldOffset(0x002C)]
-    public readonly IntPtr GameContext;
+    public readonly int GameContext;
 
     [FieldOffset(0x0044)]
-    public readonly IntPtr UserContext;
+    public readonly int UserContext;
 }
